taskflow_ai/requirements.txt
torch
transformers
gitpython
pydantic
jsonschema
datasets
accelerate
python-dotenv